add final project files
I completed the implementation of a Vulkan renderer with shaders. There is a texture and .obj model importer, which right now is only modifiable within the VS solution. I'll probably continue to work on this to make it easier via an in app GUI. There is no physics implementation, input option, or any game-like elements at the present moment as I didn't have enough time to get around to messing with implementing those features. I'm satisfied with the project as a whole, even if it might not have been as exciting as some have expected. I still really enjoyed learning more about computers graphics and the Vulkan graphics API.

As a side note, if you're don't have windows as a usable OS you won't be able to run the executable file located in the /x64/release directory in the root folder. I definitely plan on updating it in the coming months as I won't have to deal with college applications.